CustomPrimitiveTypeAttribute moved to DotVVM.Core package
