+++ src/Framework/Core/ViewModel/CustomPrimitiveTypeAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DotVVM.Framework.Configuration;
+namespace DotVVM.Framework.ViewModel;
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
 public class CustomPrimitiveTypeAttribute : Attribute
